AwsInfrastructureService/requirements.txt
-e .

aws-cdk.core
aws-cdk.aws_apigateway
aws-cdk.aws_lambda
boto3
botocore